Finished basic integration test

+++ Backend/IntegrationTests/dataRepresentation_test.go
@@ -1,7 +1,7 @@
 package IntegrationTests
 
 type league struct {
-	Id         int
+	Id         float64
 	Name       string
 	PublicView bool
 	PublicJoin bool
@@ -9,12 +9,12 @@ type league struct {
 }
 
 type team struct {
-	Id int
-	LeagueId int
+	Id float64
+	LeagueId float64
 	Name string
 	Tag string
-	Wins int
-	Losses int
+	Wins float64
+	Losses float64
 	Users []user
 	Players []player
 }
@@ -26,21 +26,21 @@ type user struct {
 }
 
 type player struct {
-	Id int
-	TeamId int
+	Id float64
+	TeamId float64
 	GameIdentifier string
 	Name string
 	mainRoster bool
 }
 
 type game struct {
-	Id int
-	LeagueId int
-	Team1Id int
-	Team2Id int
-	GameTime int
+	Id float64
+	LeagueId float64
+	Team1Id float64
+	Team2Id float64
+	GameTime float64
 	Complete bool
-	WinnerId int
-	ScoreTeam1 int
-	ScoreTeam2 int
+	WinnerId float64
+	ScoreTeam1 float64
+	ScoreTeam2 float64
 }
